+++ src/component/Pages/Home/Home.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Home = () => {
+  return (
+    <div>
+      <h2>Welcome to our page!</h2>
+    </div>
+  );
+}
+
+export default Home
