+++ src/component/Pages/NavBar/NavBar.js
@@ -1,3 +1,21 @@
-const NavBar = () => <h1>Math Magicians</h1>;
+import { Link } from 'react-router-dom';
+import '../../styles/NavBar.css';
+
+const NavBar = () => (
+  <>
+    <div className="nav">
+      <h1>Math Magicians</h1>
+      <Link className="links" to="/">
+        Home
+      </Link>
+      <Link className="links" to="/calculator">
+        Calculator
+      </Link>
+      <Link className="links" to="/Quotes">
+        Quotes
+      </Link>
+    </div>
+  </>
+);
 
 export default NavBar;
