+++ src/component/styles/NavBar.css
@@ -1,0 +1,17 @@
+.nav {
+  display: flex;
+  align-items: center;
+}
+
+.nav h1 {
+  flex-grow: 2;
+  margin-left: 20px;
+}
+
+.links {
+  margin-right: 20px;
+  text-decoration: none;
+  font-size: 20px;
+  font-weight: bold;
+  color: rgb(6, 64, 151);
+}
